Big update for 2013: More organized unit tests that can be better filtered as we make less changes in base libs and work more actively higher in the stack.
oRCVer -> oVerPatch moves a script-based .rc file that doesn't allow the user to modify it through normal means to the new post-link stamping of an exe with a P4 revision.

New FreeImage update.

More oGPU progress and of course tons of misc fixes and scrubbings.

+++ Ouroboros/External/FreeImage/Source/OpenEXR/IlmImf/ImfRgbaFile.cpp
@@ -175,11 +175,11 @@ cachePadding (ptrdiff_t size)
     while ((size >> i) > 1)
 	++i;
 
-    if (size > ((size_t)1 << (i + 1)) - 64)
-	return 64 + (((size_t)1 << (i + 1)) - size);
+    if (size > (1 << (i + 1)) - 64)
+	return 64 + (((ptrdiff_t)1 << (i + 1)) - size);
 
-    if (size < ((size_t)1 << i) + 64)
-	return 64 + (((size_t)1 << i) - size);
+    if (size < (1 << i) + 64)
+	return 64 + (((ptrdiff_t)1 << i) - size);
 
     return 0;
 }
